[Add] DeletePlaylistTest - a_user_cannot_delete_other_users_playlists

+++ app/Http/Controllers/PlaylistController.php
@@ -28,6 +28,8 @@ class PlaylistController extends Controller
 
     public function destroy(Playlist $playlist)
     {
+        $this->authorize('delete', $playlist);
+
         $playlist->delete();
 
         return response()->json([], 204);
